새로운 몬스터 추가
[/] 서버에서 몬스터 정보를 보낼 때 어떤 종류의 몬스터인지도 보내도록 변경
[/] class GameObject::m_isDeleted -> m_isValid로 변경
[/] 몇몇 코드 정리

+++ 소스/Client/Resource/Shader/common.hlsl
@@ -38,8 +38,8 @@ cbuffer cbScene : register(b3)
 cbuffer cbGameFramework : register(b4)
 {
 	float g_deltaTime;
-	float g_screenWidth;
-	float g_screenHeight;
+	uint g_screenWidth;
+	uint g_screenHeight;
 }
 
 struct VS_INPUT
